Fixed signedness conversion issue

+++ src/conference/session/media-session.cpp
@@ -4389,7 +4389,7 @@ LinphoneStatus MediaSession::sendDtmfs (const std::string &dtmfs) {
 		return -2;
 	}
 	if (!dtmfs.empty()) {
-		unsigned int delayMs = lp_config_get_int(linphone_core_get_config(getCore()->getCCore()), "net", "dtmf_delay_ms", 200);
+		unsigned int delayMs = (unsigned int) lp_config_get_int(linphone_core_get_config(getCore()->getCCore()), "net", "dtmf_delay_ms", 200);
 		d->dtmfSequence = dtmfs;
 		d->dtmfTimer = getCore()->getCCore()->sal->create_timer(MediaSessionPrivate::sendDtmf, this, delayMs, "DTMF sequence timer");
 	}
